requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pandas>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
google-genai>=1.0.0
